chore: refactor hook
